% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.R
\name{tradingview_data_from_json_string}
\alias{tradingview_data_from_json_string}
\title{Download data from the tradingView site with the passed json string}
\usage{
tradingview_data_from_json_string(...)
}
\arguments{
\item{json_string}{json string from inspect view at the bottom of TradingView}
}
\description{
Download data from the tradingView site with the passed json string
}
